[Feat] [Done] [Bishwajeet] Add Swagger documentation for challenge options API
