--- langsuit/app/api/challengeOptions/[challengeOptionsId]/route.ts
@@ -1,54 +1,0 @@
-import { eq } from "drizzle-orm";
-import { NextResponse, type NextRequest } from "next/server";
-
-import db from "@/db/drizzle";
-import { challengeOptions } from "@/db/schema";
-import { getIsAdmin } from "@/lib/admin";
-
-export const GET = async (
-  _req: NextRequest,
-  { params }: { params: { challengeOptionId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const data = await db.query.challengeOptions.findFirst({
-    where: eq(challengeOptions.id, params.challengeOptionId),
-  });
-
-  return NextResponse.json(data);
-};
-
-export const PUT = async (
-  req: NextRequest,
-  { params }: { params: { challengeOptionId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const body = (await req.json()) as typeof challengeOptions.$inferSelect;
-  const data = await db
-    .update(challengeOptions)
-    .set({
-      ...body,
-    })
-    .where(eq(challengeOptions.id, params.challengeOptionId))
-    .returning();
-
-  return NextResponse.json(data[0]);
-};
-
-export const DELETE = async (
-  _req: NextRequest,
-  { params }: { params: { challengeOptionId: number } }
-) => {
-  const isAdmin = getIsAdmin();
-  if (!isAdmin) return new NextResponse("Unauthorized.", { status: 401 });
-
-  const data = await db
-    .delete(challengeOptions)
-    .where(eq(challengeOptions.id, params.challengeOptionId))
-    .returning();
-
-  return NextResponse.json(data[0]);
-};
